requirements.txt
streamlit==1.17.0
pandas==1.3.3
numpy==1.21.2
Pillow==8.2.0
opencv-python==4.5.3.56
pydeck==0.6.2
imageio==2.9.0
requests==2.26.0
geocoder==1.38.1

